Finish Homework 6
Completed homework 6. I completed the template function to extract the time, position components, and velocity components of a given galaxy at a series of snapshot numbers. The snapshot of a given galaxy is a datafile containing information about its particles at a given time. I then found and plotted the separation distance and relative velocity magnitudes of M31 to each other galaxy as a function of time, answering questions about the number of close approaches between the MW and M31, the time scale of the merger, and the characteristic decay time of the orbit of M33.