Finish HW 7 and add Lab 8
Finished HW 7, using a symplectic leapfrog integration scheme to analytically determine the orbit of M33 around M31. This is done by assuming a Hernquist profile for the bulge and disk components of M31 and a Miyamoto-Nagai 1975 profile for its disk. These potentials are used to compute the acceleration vector on the M33 center of mass, which allows for a stepwise integration. The results of this integration are compared with the N-body simulation used throughout the course.