Upload Research Assignment 3
Add first attempt at completing galaxy project as described in header of Jupyter notebook. The goal is to select Sun-like stars in the disk of M31 and M33. A movie was then made my plotting disk particles and sun-like candidates over time. Due to time constraints, the full movie could not be generated, but a partial movie containing approximately the first quarter of the simulation data is included.