Finish lab 9
Finished lab 9, using a leapfrog numerical integration scheme to analytically determine the orbit of M33 around M31. In this notebook, we include the effects of dynamical friction and a "fudge factor" is used within the Coulomb logarithm to match the simulated orbit as closely as possible. I discuss the factors that cause the integrated orbit to eventually diverge from the simulated one.