Finish Research Assignment 5
Finished research assignment 5, the preparation for an in-class code check in. I polished the existing code, removing the previous condition for sun-like particles to be constrained by z-position and added a selection criterion based on out-of-plane velocity. Plot 1 will be a frame from the visualization made, which shows sun-like particles in the disk of M31 and M33, centered on the frame of the M31 center of mass. Finally, an additional section was added, which is a series of histograms showing the number of sunlike particles at each distance from the M31 COM over time. Plots from the first 100 snaps were generated, which will be carried out over all remaining snaps for the final result.