Finish Lab 12
Finish lab 12, regarding the temperature evolution of the universe. This lab involved adding an additional set of methods to the previously developed class containing a variety of cosmology tools. One method returns the temperature of the Universe at a given redshift. An additional function returns the "sound horizon" of the Universe at a given redshift.